+++ Bibbi.xcworkspace/xcshareddata/xcschemes/Bibbi-Workspace.xcscheme
@@ -280,62 +280,6 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "7669601E3503153E790457BC"
-               BuildableName = "Firebase_FirebaseCore.bundle"
-               BlueprintName = "Firebase_FirebaseCore"
-               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/firebase-ios-sdk/Firebase.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "C241B5678162C4E6FDA163B6"
-               BuildableName = "Firebase_FirebaseCoreInternal.bundle"
-               BlueprintName = "Firebase_FirebaseCoreInternal"
-               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/firebase-ios-sdk/Firebase.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "116920D8770A6C1C794EB0CA"
-               BuildableName = "Firebase_FirebaseInstallations.bundle"
-               BlueprintName = "Firebase_FirebaseInstallations"
-               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/firebase-ios-sdk/Firebase.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "5640303643D467DD91216895"
-               BuildableName = "Firebase_FirebaseMessaging.bundle"
-               BlueprintName = "Firebase_FirebaseMessaging"
-               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/firebase-ios-sdk/Firebase.xcodeproj">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
                BlueprintIdentifier = "4160F70E94342FB432F5ABF1"
                BuildableName = "GoogleAppMeasurementTarget.framework"
                BlueprintName = "GoogleAppMeasurementTarget"
@@ -353,6 +297,20 @@
                BlueprintIdentifier = "08D12528663D0C8DD1A99BCB"
                BuildableName = "GoogleDataTransport.framework"
                BlueprintName = "GoogleDataTransport"
+               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/GoogleDataTransport/GoogleDataTransport.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "47F627FD023B90723D5A3BC9"
+               BuildableName = "GoogleDataTransport_GoogleDataTransport.bundle"
+               BlueprintName = "GoogleDataTransport_GoogleDataTransport"
                ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/GoogleDataTransport/GoogleDataTransport.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -465,6 +423,34 @@
                BlueprintIdentifier = "3EEFCF626C2BCF3F7E6BCF80"
                BuildableName = "GoogleUtilities_UserDefaults.framework"
                BlueprintName = "GoogleUtilities-UserDefaults"
+               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/GoogleUtilities/GoogleUtilities.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "EB90FAD5F60FE50D10E84EFD"
+               BuildableName = "GoogleUtilitiesPrivacy.framework"
+               BlueprintName = "GoogleUtilitiesPrivacy"
+               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/GoogleUtilities/GoogleUtilities.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "855D7D904350B563639C3548"
+               BuildableName = "GoogleUtilities_GoogleUtilitiesPrivacy.bundle"
+               BlueprintName = "GoogleUtilities_GoogleUtilitiesPrivacy"
                ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/GoogleUtilities/GoogleUtilities.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -648,6 +634,20 @@
                BuildableName = "Mixpanel.framework"
                BlueprintName = "Mixpanel"
                ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/mixpanel-swift/Mixpanel.xcodeproj">
+            </BuildableReference>
+         </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "82B33931410C5E34C202CFA7"
+               BuildableName = "Promises_FBLPromises.bundle"
+               BlueprintName = "Promises_FBLPromises"
+               ReferencedContainer = "container:Tuist/Dependencies/SwiftPackageManager/.build/checkouts/promises/Promises.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
